Remove duplicated code blocks.
