+++ iorek-notify/src/main/java/com/stuartbeard/iorek/notify/config/CompromisedPasswordNotificationConfiguration.java
@@ -8,6 +8,7 @@ package com.stuartbeard.iorek.notify.config;
 import com.stuartbeard.iorek.notify.event.CompromisedPasswordNotifier;
 import com.stuartbeard.iorek.notify.service.CompromisedPasswordNotificationService;
 import com.stuartbeard.iorek.service.PasswordCheckingService;
+import com.stuartbeard.iorek.service.config.CommonCompromisedPasswordConfiguration;
 import com.stuartbeard.iorek.service.config.CompromisedPasswordThresholdConfigurationProperties;
 import com.stuartbeard.iorek.service.external.CompromisedPasswordService;
 import com.stuartbeard.iorek.service.recorder.PasswordCheckRecorder;
@@ -28,7 +29,10 @@ import org.springframework.context.annotation.Import;
  * @since 1.0.0
  */
 @Configuration
-@Import(PwnedPasswordsConfiguration.class)
+@Import({
+    CommonCompromisedPasswordConfiguration.class,
+    PwnedPasswordsConfiguration.class
+})
 public class CompromisedPasswordNotificationConfiguration {
 
     /**
@@ -52,29 +56,6 @@ public class CompromisedPasswordNotificationConfiguration {
     }
 
     /**
-     * Conditionally instantiates a Password Checking Service Spring Bean if one has not already been created. This is
-     * the centralised logic responsible for orchestrating the request to the external compromised password
-     * service, aggregating the result, and calculating the perceived password risk level from configuration.
-     *
-     * @param compromisedPasswordService An implementation that connects to an external datasource for checking
-     *                                   passwords for <strong>known</strong> compromise and returns the count
-     *                                   associated with their appearances in known data breaches.
-     * @param credentialSafetyConfig     Configuration properties that set the thresholds for when a given password is marked as
-     *                                   <code>PasswordRiskLevel.COMPROMISED</code> or
-     *                                   <code>PasswordRiskLevel.SEVERELY_COMPROMISED</code>.
-     * @param passwordCheckRecorder      Used to record the request flow and password check result
-     *                                   (with no identifiable information) against some form of monitoring system.
-     * @return an instance of the core Password Checking Service
-     */
-    @Bean
-    @ConditionalOnMissingBean
-    public PasswordCheckingService passwordCheckingService(CompromisedPasswordService compromisedPasswordService,
-                                                           CompromisedPasswordThresholdConfigurationProperties credentialSafetyConfig,
-                                                           PasswordCheckRecorder passwordCheckRecorder) {
-        return new PasswordCheckingService(compromisedPasswordService, credentialSafetyConfig, passwordCheckRecorder);
-    }
-
-    /**
      * Provides a default implementation of the {@link CompromisedPasswordNotificationService} class (i.e. when called
      * to send a notification an {@link UnsupportedOperationException} will be thrown. Consumers of this library should
      * implement their own class that caters to how they want to notify users out of band (e.g. Email,
@@ -86,24 +67,5 @@ public class CompromisedPasswordNotificationConfiguration {
     @ConditionalOnMissingBean
     public CompromisedPasswordNotificationService compromisedPasswordNotificationService() {
         return new CompromisedPasswordNotificationService() {};
-    }
-
-    /**
-     * Conditionally instantiates a Password Check Recorder Spring Bean if one has not already been created. This is to
-     * provide a simple zero operation implementation of the interface to ensure that monitoring functionality works correctly.
-     *
-     * @return a default implementation of the Password Check Recorder interface. This implementation does nothing when called.
-     */
-    @Bean
-    @ConditionalOnMissingBean
-    public PasswordCheckRecorder passwordCheckRecorder() {
-        return new PasswordCheckRecorder() {};
-    }
-
-    @Bean
-    @ConditionalOnMissingBean
-    @ConfigurationProperties("compromised.password.thresholds")
-    public CompromisedPasswordThresholdConfigurationProperties credentialSafetyConfig() {
-        return new CompromisedPasswordThresholdConfigurationProperties();
     }
 }
